introduzione game over e vittoria - manca da sistemare la gestione dei power up, delle monete, dei nemici da metà mappa ed eventuale schermata iniziale e musica

+++ src/main/java/beans/Player.java
@@ -256,14 +256,14 @@ public class Player extends MovableGameObject {
 	/**
 	 * @return the vel_x
 	 */
-	public float getVel_x() {
+	public int getVel_x() {
 		return vel_x;
 	}
 
 	/**
 	 * @return the vel_y
 	 */
-	public float getVel_y() {
+	public int getVel_y() {
 		return vel_y;
 	}
